Renames and setup
