+++ lib/src/test/java/edu/stanford/futuredata/macrobase/analysis/summary/IntSetAsArrayTest.java
@@ -1,6 +1,6 @@
 package edu.stanford.futuredata.macrobase.analysis.summary;
 
-import edu.stanford.futuredata.macrobase.analysis.summary.util.IntSet;
+import edu.stanford.futuredata.macrobase.analysis.summary.util.IntSetAsArray;
 import org.junit.Test;
 
 import java.util.HashSet;
@@ -8,10 +8,10 @@ import java.util.Set;
 
 import static org.junit.Assert.*;
 
-public class IntSetTest {
+public class IntSetAsArrayTest {
     @Test
     public void testSimple() {
-        IntSet s1 = new IntSet(1, 2);
+        IntSetAsArray s1 = new IntSetAsArray(1, 2);
         Set<Integer> s2 = new HashSet<>(2);
         s2.add(1);
         s2.add(2);
@@ -23,12 +23,12 @@ public class IntSetTest {
 
     @Test
     public void testContains() {
-        IntSet s1 = new IntSet(1,2);
+        IntSetAsArray s1 = new IntSetAsArray(1,2);
         assertTrue(s1.contains(1));
         assertTrue(s1.contains(2));
-        IntSet s2 = new IntSet(1,2,3);
+        IntSetAsArray s2 = new IntSetAsArray(1,2,3);
         assertTrue(s2.contains(s1));
-        IntSet s3 = new IntSet(1,4);
+        IntSetAsArray s3 = new IntSetAsArray(1,4);
         assertFalse(s3.contains(s1));
     }
 }
